+++ src/components/IndividualPageContent/CollapsibleCards.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Stack, Typography, useTheme, Box, Button} from "@mui/material";
-import {grey} from "../../../../themes/colors/aptosColorPalette";
+import {grey} from "../../themes/colors/aptosColorPalette";
 
 // return true if there is at least 1 card is not expanded
 function getNotAllExpanded(expandedList: boolean[]): boolean {
