+++ src/components/IndividualPageContent/EmptyValue.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Box} from "@mui/material";
-import {grey} from "../../../../themes/colors/aptosColorPalette";
+import {grey} from "../../themes/colors/aptosColorPalette";
 
 export default function EmptyValue() {
   return <Box color={grey[450]}>N/A</Box>;
