+++ src/pages/Transaction/Tabs/BlockMetadataOverviewTab.tsx
@@ -8,8 +8,8 @@ import {
 } from "../../Transactions/helpers";
 import Row from "./Components/Row";
 import HashButton, {HashType} from "../../../components/HashButton";
-import ContentBox from "./Components/ContentBox";
-import ContentRow from "./Components/ContentRow";
+import ContentBox from "../../../components/IndividualPageContent/ContentBox";
+import ContentRow from "../../../components/IndividualPageContent/ContentRow";
 import TransactionStatus from "../../../components/TransactionStatus";
 import {useGetInDevMode} from "../../../api/hooks/useGetInDevMode";
