+++ src/pages/Transaction/Tabs/GenesisTransactionOverviewTab.tsx
@@ -3,8 +3,8 @@ import {Types} from "aptos";
 import {Stack, Box} from "@mui/material";
 import {renderGas, renderSuccess} from "../../Transactions/helpers";
 import Row from "./Components/Row";
-import ContentRow from "./Components/ContentRow";
-import ContentBox from "./Components/ContentBox";
+import ContentRow from "../../../components/IndividualPageContent/ContentRow";
+import ContentBox from "../../../components/IndividualPageContent/ContentBox";
 import TransactionStatus from "../../../components/TransactionStatus";
 import {useGetInDevMode} from "../../../api/hooks/useGetInDevMode";
