+++ src/pages/Transaction/Tabs/PayloadTab.tsx
@@ -3,8 +3,8 @@ import {Types} from "aptos";
 import {Box, Typography} from "@mui/material";
 import {renderDebug} from "../../utils";
 import {useGetInDevMode} from "../../../api/hooks/useGetInDevMode";
-import CollapsibleCard from "./Components/CollapsibleCard";
-import JsonCard from "../../../components/JsonCard";
+import CollapsibleCard from "../../../components/IndividualPageContent/CollapsibleCard";
+import JsonCard from "../../../components/IndividualPageContent/JsonCard";
 
 type PayloadTabProps = {
   transaction: Types.Transaction;
